Pingo 1.3 Update
maybe the final update
- UWP versions (win10,winphone10) are now official
- macOS version is now official (thx to my "Acer"tosh laptop)
- gamepad support
- saving bugs fixed
- dogs removed

+++ objects/oSaveSystem/Create_0.gml
@@ -1,17 +1,22 @@
 /// @description init
-if (file_exists("PingoSaveData/Pingo.sav"))
+
+if !(os_browser != browser_not_a_browser) && (file_exists("PingoSaveData/Pingo.sav"))
 {
 	show_debug_message("OLD SAVEDATA WAS FOUND! COPYING...");
+	directory_create("default");
 	file_copy("PingoSaveData/Pingo.sav","default/Pingo.sav");
 	file_delete("PingoSaveData/Pingo.sav");
 	directory_destroy("PingoSaveData");
 	show_debug_message("DONE COPYING SAVEDATA!");
 }
 
+image_speed = 0;
 global.loadid = -1;
 global.saveid = -1;
+global.levelid = -1;
 global.loadbuff = -1;
 global.savebuff = -1;
+global.levelbuff = -1;
 global.CurrentLevel = 1;
 global.CurrentWorld = 1;
 
@@ -31,7 +36,5 @@ for (var i = 1; i < 31; i++)
 	global.StarCount[1,i] = 0;
 	global.StarCount[2,i] = 0;
 }
-
-PerformEvent = true;
 
 scrLoad();
